requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
automatic-speech-recognition==1.0.4
backcall==0.2.0
beautifulsoup4==4.9.3
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bleach==3.2.2
Brotli==1.0.9
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
docx==0.2.4
entrypoints==0.3
enum34==1.1.10
et-xmlfile==1.0.1
filelock==3.0.12
Flask==2.0.1
Flask-Compress==1.9.0
Flask-Cors==3.0.10
Flask-JSON==0.3.4
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.3.3
google-api-core==1.26.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
GPUtil==1.4.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.32.0
h5py==2.10.0
huggingface-hub==0.0.8
idna==2.6
imageio==2.9.0
img2pdf==0.4.0
imgaug==0.4.0
install==1.3.4
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
Js2Py==0.70
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lesscpy==0.14.0
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.3
mistune==0.8.4
mogrify==0.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
nltk==3.6.2
notebook==6.2.0
numexpr==2.7.3
numpy==1.20.3
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pdf2image==1.14.0
pickleshare==0.7.5
pikepdf==2.5.2
Pillow==8.1.0
pipwin==0.5.1
ply==3.11
pocketsphinx @ file:///C:/Users/Bhavesh/pipwin/pocketsphinx-0.1.15-cp38-cp38-win_amd64.whl
prometheus-client==0.9.0
prompt-toolkit==3.0.11
protobuf==3.14.0
psycopg2==2.8.6
pvporcupine==1.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Bhavesh/pipwin/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.7.4
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pyrsistent==0.17.3
pySmartDL==1.3.4
pytesseract==0.3.7
python-dateutil==2.8.1
python-docx==0.8.10
python-speech-features==0.6
pytz==2020.5
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4
pyzmq==21.0.1
qtconsole==5.0.1
QtPy==1.9.0
regex==2021.4.4
requests==2.18.4
requests-oauthlib==1.3.0
rsa==4.7
sacremoses==0.0.45
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
soupsieve==2.2
SpeechRecognition==3.8.1
SQLAlchemy==1.4.18
table-ocr==0.2.5
tables==3.6.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
tesseract==0.1.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.1.14
tokenizers==0.10.3
torch==1.8.1+cpu
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
transformers==4.6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.22
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
